Relocated delete.py to ssh
