+++ arc/job/ssh.py
@@ -10,6 +10,7 @@ import logging
 import os
 import time
 import datetime
+import re
 
 import paramiko
 
@@ -301,5 +302,37 @@ def check_job_status_in_stdout(job_id, stdout, server):
             return 'errored on node ' + status_line.split()[-1][-2:]
         else:
             raise ValueError('Unknown cluster software {0}'.format(servers[server]['cluster_soft']))
+
+
+def delete_all_arc_jobs(server_list):
+    """
+    Delete all ARC-spawned jobs (with job name starting with `a` and a digit) from :list:servers
+    (`servers` could also be a string of one server name)
+    Make sure you know what you're doing, so unrelated jobs won't be deleted...
+    Useful when terminating ARC while some (ghost) jobs are still running.
+
+    Args:
+        server_list (list): List of servers to delete ARC jobs from.
+    """
+    if isinstance(server_list, str):
+        server_list = [server_list]
+    for server in server_list:
+        print('\nDeleting all ARC jobs from {0}...'.format(server))
+        cmd = check_status_command[servers[server]['cluster_soft']] + ' -u ' + servers[server]['un']
+        ssh = SSHClient(server)
+        stdout = ssh.send_command_to_server(cmd)[0]
+        for status_line in stdout:
+            s = re.search(r' a\d+', status_line)
+            if s is not None:
+                if servers[server]['cluster_soft'].lower() == 'slurm':
+                    job_id = s.group()[1:]
+                    server_job_id = status_line.split()[0]
+                    ssh.delete_job(server_job_id)
+                    print('deleted job {0} ({1} on server)'.format(job_id, server_job_id))
+                elif servers[server]['cluster_soft'].lower() == 'oge':
+                    job_id = s.group()[1:]
+                    ssh.delete_job(job_id)
+                    print('deleted job {0}'.format(job_id))
+    print('\ndone.')
 
 # TODO: delete scratch files of a failed job: ssh nodeXX; rm scratch/dhdhdhd/job_number
