+++ arc/utils/delete.py
@@ -7,7 +7,7 @@ ARC - Automatic Rate Calculator
 
 from __future__ import (absolute_import, division, print_function, unicode_literals)
 
-from arc.main import delete_all_arc_jobs
+from arc.job.ssh import delete_all_arc_jobs
 from arc.settings import servers
 
 ################################################################################
